Move all local executors to python executors module
